moved inspection methods to a dedicated script and refactored
